Add civiform container healthcheck, wait until service is healthy. (#97)
* Add civiform container healthcheck, wait until service is healthy.

By adding container healthchecks to civiform task definition, we can
tell if the service is in the desired state by inspecting the service
deployments.

Add code in bin/{setup,deploy} to wait for the service to reach desired
state. Containers have to pass their healthchecks for the deployment to
complete so this solves the "deploy tool completes without any error but
the civiform container is crashlooping" issue.