Reading configs directly from .sh rather than env variables (#134)
Reading configs directly from .sh rather than env variables

More information here:
https://docs.google.com/document/d/1M27DF1IJypaeyLQrTZAy-y2DEszUV-EgGkhuNgOMzRE/edit#heading=h.bknox3ib8wpb

Avritt's change touches some of the same files (config loader). As we
are both on and off this week, I suggest that whoever is ready to submit
first submits and the other does the merge, so that we both stay
unblocked. I had a look and the merge looks similarly feasible no matter
which change goes in first.

---------

Co-authored-by: Jessica Hummel <jhummel@google.com>