Add the plumbing for enabling variable auto generation via env var docs (#147)
This PR adds a new script "run" which runs before run.py.
The purpose is to download and install the env-var-docs python package.
We create a python virtual environment(venv), install the dynamically
downloaded env-var-docs package in it and run our python deploy code
inside the venv so it has access to env-var-docs.

#132 contains the feature this would enable, whose longer term goal is
to automatically generate the variables that allow configuring the
server via the deploy system.
Note that this PR does not yet enable the feature. It merely provides
the script that can do so.
When we are happy to enable the feature, we should replace all calls to
run.py with run, which then calls run.py in turn.

Some reasoning about the location of this: 
We need to enable the python packages across all deployment implement
system (dockerized deployment, checkout deployment, AWS. Azure). The
run.py script is the first step that is shared between all those
options. There are some other benefits to placing it here rather than
somewhere in the deploy repository (e.g. you can use run during
development to not have to code across multiple branches)
We can not enable the package in run.py directly, because the activation
of the python virtual environment needs to happen in the parent process
which the python process is started from before the code is started.

---------

Co-authored-by: Jessica Hummel <jhummel@google.com>