bump tries on server start poll and edit comment (#168)
The server start checks have been timing out periodically even when the
service is healthy. This can lead to confusion with people thinking
their deploy failed when it actually succeeded. This PR bumps the number
of times the check will run to 30, which gives a longer runway for the
service to start, and also adds a more detailed failure message to let
the user know their service might have started even if the checks time
out.