Optimizations for our ecs_fargate_service
This cleans up a bunch of resources to minimize the amount of code we're using, such as removing 'count' and 'for_each' when we only have one of something, and moving some of the values we were defining as an input to the module from app.tf into the module itself, since we don't really need to expose them to users to change them.

Additionally, this renames name_prefix to app_prefix and only the app_prefix is passed in. However, the module will create a local named name_prefix and append -civiform to it. This is to reduce the number of resources that get destroyed and recreated just due to a name change, and we need to use the shorted app_prefix for the loadbalancer target resource since it has a maximum character limit of 32.

Finally, this puts tags back to how they were and adds 'moved' blocks. This allows Terraform to know that certain resources have changed namespaces so it doesn't try to recreate them. This should minimize the amount of changes TF makes when doing a bin/deploy and prevents us from needing to destroy and recreate the load balancer. Note that the LB resource name was changed because TF was getting confused with this particular resource when it was still aws_lb.lb for some reason.