Turn on the service deployment circuit breaker
By turning on the circuit breaker, we prevent a deployment with the
service crash-looping from repeatedly pulling images from DockerHub,
which can cause rate limiting. The deployment will attempt to start the
service 10 times. After failing 10 times, it will stop trying and mark
the deployment as failed.

Additionally, we will now roll back to the previous, working deployment
if the new deployment fails. Some logic has been added to detect when
this happens, as it will simply appear that the PRIMARY deployment is
COMPLETE, and we need to check the deployment ID to detect it's the old
deployment.

This also modifies the aws_cli script to detect when it has failed and
stops polling the state of the service and prints information on how to
check the logs.