Remove detection of current deployment in deploy.py
Checking the ID of the deployment we are deploying now vs. the ID of the PRIMARY COMPLETED deployment at the end of the deployment is sufficient to detect if we've rolled back to a previous deployment because this one failed. This removes the additional complexity around checking the ID of the current deployment before doing the terraform apply and passing it into the wait_for_ecs_service_healthy function.