Add ability to force unlock TF state
Sometimes, if a failure occurs or you've Ctrl+C'ed out of the middle of
an apply, the Terraform state may be left in a locked state. When this
happens, you must run 'terraform force-unlock' with the lock ID to
unlock it.

This change detects when an apply results in a lock error, and prompts
the user to rerun the command with the --force-unlock flag. It does this
by capturing the stderr stream of the 'terraform apply' command to
inspect the message if an error occurs. An added bonus is that since we
are using Popen to execute the command, we can handle things more
gracefully if a user Ctrl+Cs out of the apply and allow Terraform to
release the lock, reducing the amount of ways we can get into this
failure state.

In order to support the --force-unlock flag, there will be a change to
the civiform-deploy repo as well.