Perform fixes during the run if running interactively
If we hit a TF lock state issue, or the S3 digest issue, this allows the
user to fix the problem during this particular invocation of the
command, so they do not need to rerun the command. If we are not running
in a TTY, we'll fall back to prompting the user to rerun the command
with the appropriate flag.