Detect if backend resources already exist and optionally destroy
If the user did a failed bin/setup, or canceled out of bin/setup early,
we may be left with the S3 bucket and DynamoDB table used to maintain
the Terraform backend state. This detects if those resources exist, and
gives the user the option to clean them up, warning them that if they do
this when there are actual deployment resources applied, we'll lose
track of them and cause more problems.