Upgrade to PostgreSQL 16
This adds the machinery to upgrade a CiviForm instance to use PostgreSQL
16. All current versions should be on version of 12, with most probably
on 12.17 or 12.18. This checks the currently deployed version, and
verifies that it is sufficiently new enough to make the jump to 16. It
gives the user information about how the upgrade will go, and then
requires the user to set ALLOW_POSTGRESQL_UPGRADE before actually
performing the upgrade to ensure the user has read all the necessary
warnings.