Create dumpdb command
This adds a new command called "dumpdb" and a new Terraform module
called dbaccess. The module creates a temporary EC2 host with a public
IP, with a security group that allows SSH access only to the host this
script is running from, and with PostgreSQL access to the RDS database.
The script sets these resources up, along with a temporary keypair,
SSHs into the temporary host to install postgresql-client and run
pg_dump, SCPs the dump file to the host running this script, then cleans
up all the resources.  This also refactors the CIDR state machine into a
separate file. While this command does not use the whole state machine,
it uses the function to detect the public IP of the current host.