Update nuke.yaml
This updates the filter list to prevent deleting important things. Note that not all of these things are needed for the e2e-test, but since we're using this file as a template for our staging instances now, it includes thing like the deploy action role and makes some of the filters more generic.