Fix upgrade paths for different postgres versions
It would seem that RDS will let 12.17 ONLY upgrade to 16.1, and 12.18
ONLY upgrade to 16.2, not 16.1. This means we have to maintain a table
of valid upgrade paths here until everyone has made it onto 16, at which
point RDS will upgrade minor versions automatically.

This changes the deploy logic, making it a bit less generic in order to
handle this additional complexity of specific major.minor valid upgrade
paths.