Only apply database changes immediately when doing a major upgrade
Because we had the 'apply_immediately' parameter on the aws_db_instance hard-coded to true, this meant that whenever there was a new postgres minor available in AWS, the deployment would force the database to upgrade to that minor during the deployment, incurring downtime. Since CEs would prefer to have that downtime happen during the maintenance window instead since this happens in off hours, this only forces the parameter to true during a major database upgrade, where we are also changing the database parameters and need to make both changes at the same time.

If a user wishes, they can set APPLY_DATABASE_CHANGES_IMMEDIATELY=true in their config to override this behavior.