Fix restoring from a snapshot
When we restore from a snapshot, Terraform applies the temporary
database password to the restored database, rather than using the actual
database password as applied after the Terraform setup finishes (we do
this so we don't store the actual password in Terraform state). This reads the actual
password for the database from the secret and applies it when restoring.

While the code supports doing a bin/deploy with a snapshot identified
(that is, destroying your current database and restoring from a snapshot
in-place), we've decided to disable this code path for now until we get
a more imperative system for restoring from a snapshot in place in order
to avoid accidental data loss.

In a workflow where you want to set up a separate instance with a
different app_prefix, but based on a snapshot from the old app_prefix,
during bin/setup, this prompts the user to load the secret value from the previous
app_prefix, or to enter the password manually.

Additionally, when we are standing up a brand new CiviForm instance
based on a snapshot from another deployment, that snapshot is encrypted
with the KMS key from the old deployment. This results in Terraform
wanting to destroy the database and recreate it again on the next
deploy. Instead, we now copy the snapshot and re-encrypt it with the new
instance's key, then deploy the new database from that snapshot. A
subsequent deploy with POSTGRES_RESTORE_SNAPSHOT_IDENTIFIER removed will
remove that copy.

Finally, this updates ecs-service-autoscaling to 1.0.10 to squash a bug
that caused it to always add 'null' into alarm action lists.