Replace colon with hypen in snapshot copy name
Automated RDS snapshots are prefixed with "rds:". However, only those
automated snapshots are allowed to be called this, so when we create the
copy of the snapshot for use in a new deployment, replace the colon with
a hyphen.

Because AWS is, shall we say, "special", it can only contain letters,
digits and hyphens, and can't contain two consecutive hyphens or end in a
hyphen for reasons surpassing understanding. We don't guard against that
here since the source snapshot name has to be created with those
restrictions as well.