Change default LB policy
This sets the default Load Balancer policy to
ELBSecurityPolicy-TLS13-1-2-2021-06, which allows TLSv1.3 and 1.2. This
is the recommended policy. See
https://docs.aws.amazon.com/elasticloadbalancing/latest/application/describe-ssl-policies.html
for details on all the policies.

In fact, clients were still able to connect via TLSv1.3 due to changes
in the ALB applied last year. https://aws.amazon.com/about-aws/whats-new/2023/03/application-load-balancer-tls-1-3/

This policy ensures that connections between the ALB and the server is
also TLSv1.3 or 1.2.  This is somewhat less important since it is all
internal to the VPC, but a good idea nonetheless.

This also updates the ALB policy when setting up pgadmin.

Finally, this exposes the security policy as a variable so it can be
overriden if needed.